Improve duration format in entity card
